Struct kaleidoscope::Token

pub struct Token {
    pub token_kind: TokenKind,
    pub span: String,
    pub start: FileIndex,
    pub end: FileIndex,
}

A token in a Kaleidoscope file.

This structural representation of a token contains the possible TokenKind of the token, the token as a string (stored as span), as well as the start and end indices of the token.

Fields
token_kind: TokenKindspan: Stringstart: FileIndexend: FileIndex
Implementations

impl Token

pub fn new(
token_kind: TokenKind,
span: String,
start: FileIndex,
end: FileIndex
) -> Token

Generate a new token from known values.

pub fn new_eof(index: FileIndex) -> Token

A new token whose TokenKind is TokenKind::Eof.

pub fn is_eof(&self) -> bool

Check if this token represents an EOF character.

Example

use kaleidoscope_lexer::token::{FileIndex, Token};

assert!(Token::new_eof(FileIndex::new(None, 0)).is_eof());

pub fn is_terminating(&self) -> bool

Check whether this token kind represents a sentinel value that tells the parser to halt and return whatever it has processed. Such tokens include semicolons and EOFs.

pub fn denotes_end_of_expression(&self) -> bool

Check whether this token is allowed to occur at the end of an expression.

pub fn is_empty(&self) -> bool

Check if the span in the token is empty or not.

pub fn add_unit(
&mut self,
unit: char,
index: FileIndex
) -> Result<bool, Error<ErrorKind>>

Add a character into the token. This method returns a boolean value if no error occurs.

If true is returned, this means that unit is a character that this token cannot accept, so this token is complete. If false is returned, it means that unit has been added to the token’s span and can continue to collect more characters.

pub fn resolve(&mut self, index: FileIndex) -> Result<bool, Error<ErrorKind>>

Resolve the value of the token when the end of the sequence happens. If this function is successful, it will always return Ok(true).

pub fn borrow_span(&self) -> &str

Borrow the span occupied by the token as a string slice.